requirements.txt
# ============================================================
# requirements.txt — Trading Bot Dependencies
# pip install -r requirements.txt
# ============================================================

# --- Core Trading ---
ccxt==4.2.25
python-dotenv==1.0.1
pyyaml==6.0.1

# --- Telegram Bot ---
python-telegram-bot==20.7

# --- Data Processing ---
pandas==2.2.0
numpy==1.26.3
# ta-lib kurulumu: https://github.com/TA-Lib/ta-lib-python
# Windows: pip install TA_Lib-0.4.28-cp310-cp310-win_amd64.whl
ta-lib==0.4.28

# --- AI / LLM ---
openai==1.12.0
google-generativeai==0.3.2
aiohttp==3.9.3
requests==2.31.0

# --- Async & Scheduling ---
APScheduler==3.10.4

# --- Optional: Caching ---
redis==5.0.1

# --- Monitoring ---
psutil==5.9.8

# --- Dev & Testing ---
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
black==24.2.0
flake8==7.0.0
